implement initial trust checking
When a key was never seen before we check the signatures. If there is
at least one trusted signature we trust the key. If a key is
different from a known key in this repository we consider the
repository as not trusted.

This also refactors the init interface in the sense that we no longer
have to run it from an empty repository that will then get
initialized and we have to push this manually to the remote. With the
new interface the URL now has to be provided as the first positional
argument and the implementation will do all the magic, no matter
where it got invoked from.