add an initial version of GitHub Actions testing
This is also adapting the code in some aspects to make it compatible
with older versions of Python and pygit2. Those adaptions are in
particular:
- Older versions of pygit2 didn't have enums FileMode and ObjectType. We
  define the needed values if they are missing.
- Older versions of Python didn't provide default values for the
  to_bytes function. We provide the values ourselves now.
- Older versions of pygit2 didn't convert object ids to strings
  implicitly. We do this explicitly now wherever needed.
- Older versions of Python were less robust in parsing quotes nested in
  f-string literals. Work around those constructs that found to be
  problematic.

The Windows build is currently disabled since there is still an issue
with the remote helper that I need to debug.

Ubuntuo 20.04 on ARM is also disbled because the machine currently does
not seem to start up on GitHub.